Generate commonjs modules.
This allows us to:
* Import haskell files from javascript
* Hot reload modules

See the demo of both of these features: http://youtu.be/DyFcrJL6kxk

CommonJS is important because:
* It's the state of the art in js module systems.
* It allows us to link in many haskell modules sharing the same runtime.

This prototype uses webpack in place of GHC as the linker. I chose
webpack because:
* It's fast
* It supports custom loaders
  ([haskell-loader](https://github.com/joelburget/haskell-loader))
* It supports hot module replacement

Other notes:

Currently everything in the runtime implemented in js is dumped onto
`window` (see `lib/runtime.js`). I don't like this at all, but it should
be easy to fix and is equivalent to what Haste already does.

`hastec` is broken in this commit. Easily fixed.

Hot module replacement support is a bit of a hack. If you need to run an
action on module loading or unloading, that action is exported as
`onHotLoad` or `onHotUnload`, which is detected and wired in by
`haskell-loader`

I've done previous work on using CommonJS modules from Haskell
(https://github.com/joelburget/haste-commonjs).